move some classes from manager.py to separate file
